move GenericCounter into counters.tsx
