+++ app/javascript/mastodon/components/counters.tsx
@@ -60,6 +60,17 @@ export const GenericCounterRenderer: React.FC<GenericCounterRendererProps> = ({
   }
 };
 
+interface GenericCounterProps {
+  value: number;
+  children?: never;
+}
+
+const _GenericCounter: React.FC<GenericCounterProps> = ({ value }) => {
+  const shortNumber = toShortNumber(value);
+  return <GenericCounterRenderer value={shortNumber} />;
+};
+export const GenericCounter = memo(_GenericCounter);
+
 interface StatusesCounterProps {
   value: number;
   children?: never;
