+++ app/javascript/mastodon/components/server_banner.jsx
@@ -8,8 +8,8 @@ import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import { fetchServer } from 'mastodon/actions/server';
+import { GenericCounter } from 'mastodon/components/counters';
 import { ServerHeroImage } from 'mastodon/components/server_hero_image';
-import { GenericCounter } from 'mastodon/components/short_number';
 import { Skeleton } from 'mastodon/components/skeleton';
 import Account from 'mastodon/containers/account_container';
 import { domain } from 'mastodon/initial_state';
